+++ packages/core/src/components/SingleSelect/FlatVariant/index.ts
@@ -1,0 +1,1 @@
+export { FlatVariant as default } from './FlatVariant';