+++ packages/core/src/components/SingleSelect/Options/Option/types.ts
@@ -3,12 +3,14 @@ import { Option } from '../../types';
 
 export interface OptionStyledProps extends HTMLProps<HTMLLIElement>, WithThemeProp {
     selected: boolean;
+    variant: 'outlined' | 'filled' | 'flat';
     hasError?: boolean;
     hovered?: boolean;
 }
 
 export interface OptionProps extends Option, Omit<HTMLProps<HTMLLIElement>, 'onClick' | 'value' | 'label' | 'size'>, WithThemeProp {
     size?: 'S' | 'M';
+    variant: 'outlined' | 'filled' | 'flat';
     onClick: (op: Option) => void;
     hasError?: boolean;
 }