+++ packages/core/src/components/SingleSelect/Options/types.ts
@@ -5,7 +5,7 @@ export interface OptionsProps extends Omit<HTMLProps<HTMLUListElement>, 'size'> 
     id?: string;
     size?: 'S' | 'M';
     options: Option[];
-    variant: 'outlined' | 'filled';
+    variant: 'outlined' | 'filled' | 'flat';
     isNested?: boolean;
     hasError?: boolean;
     onOptionClick: (option: Option | React.MouseEvent<HTMLLIElement>) => void;
@@ -14,5 +14,5 @@ export interface OptionsProps extends Omit<HTMLProps<HTMLUListElement>, 'size'> 
 export interface StyledOptionsProps extends WithThemeProp {
     size?: 'S' | 'M';
     isNested?: boolean;
-    variant: 'outlined' | 'filled';
+    variant: 'outlined' | 'filled' | 'flat';
 }