+++ packages/core/src/components/SingleSelect/SingleSelect.stories.tsx
@@ -2,7 +2,7 @@ import { defaultTheme, SingleSelectTheme } from '@medly-components/theme';
 import React from 'react';
 import { Option, SelectProps } from './types';
 
-export const variants: SelectProps['variant'][] = ['outlined', 'filled'];
+export const variants: SelectProps['variant'][] = ['outlined', 'filled', 'flat'];
 export const sizes: SelectProps['size'][] = ['S', 'M'];
 
 export const options = [