+++ packages/layout/src/components/SideNav/NavItem/NavItem.styled.tsx
@@ -4,15 +4,9 @@ import { defaultTheme } from '@medly-components/theme';
 import { css, getFontStyle, styled } from '@medly-components/utils';
 import { NavItemStyledProps } from './types';
 
-const getStyle = ({
-    navItem,
-    state,
-    isExpanded,
-    isHovered
-}: NavItemStyledProps & { state: 'default' | 'hovered' | 'pressed' | 'active' }) => css`
-    background-color: ${(isExpanded || isHovered) && navItem.bgColor[state]};
+const getStyle = ({ navItem, state }: NavItemStyledProps & { state: 'default' | 'hovered' | 'pressed' | 'active' }) => css`
+    background-color: ${navItem.bgColor[state]};
     ${SvgIcon} {
-        background-color: ${!isExpanded && !isHovered && navItem.bgColor[state]};
         * {
             fill: ${navItem.icon.color[state]};
         }
@@ -25,12 +19,12 @@ export const NavItemStyled = styled('li').attrs(({ theme: { sideNav } }) => ({ .
     text-decoration: none;
     box-sizing: border-box;
     display: grid;
-    margin-left: 1.2rem;
+    margin: 0 1.2rem;
     border-radius: ${({ navItem }) => navItem.borderRadius};
+    box-shadow: ${({ theme, isActive }) => isActive && `0 0 2.4rem ${theme.colors.grey[100]}`};
     grid-template-columns: ${({ openSize, closeSize }) => `calc(${closeSize} - 1.2rem) calc(${openSize} - ${closeSize} - 1.2rem) `};
     justify-items: flex-start;
     align-items: center;
-    overflow: hidden;
     user-select: none;
     min-height: ${({ navItem }) => navItem.minHeight};
     transition: all 100ms ease-out;
@@ -47,9 +41,11 @@ export const NavItemStyled = styled('li').attrs(({ theme: { sideNav } }) => ({ .
         text-overflow: ellipsis;
         white-space: nowrap;
         justify-self: left;
+        overflow: hidden;
         transition: opacity 200ms ease-out, color 100ms ease-out;
         opacity: ${({ isHovered, isExpanded }) => (isHovered || isExpanded ? 1 : 0)};
-        ${({ theme }) => getFontStyle({ theme, fontVariant: theme.sideNav.navItem.text.textVariant, fontWeight: 'Medium' })}
+        ${({ theme, isActive }) =>
+            getFontStyle({ theme, fontVariant: theme.sideNav.navItem.text.textVariant, fontWeight: isActive ? 'Medium' : 'Regular' })}
     }
 
     ${props => getStyle({ ...props, state: props.isActive ? 'active' : 'default' })};