+++ packages/layout/src/components/SideNav/SideNav.tsx
@@ -9,7 +9,7 @@ import ToggleSwitch from './ToggleSwitch';
 import { SideNavProps, SideNavStaticProps } from './types';
 
 export const SideNav: FC<SideNavProps> & WithStyle & SideNavStaticProps = props => {
-    const { active, children, defaultActive, onChange } = props,
+    const { active, children, defaultActive, onChange, hideShadow } = props,
         id = props.id || 'medly-sidenav';
 
     const ref = useRef(null),
@@ -34,8 +34,14 @@ export const SideNav: FC<SideNavProps> & WithStyle & SideNavStaticProps = props 
 
     return (
         <SideNavContext.Provider value={{ isHovered, isExpanded, activeItem: active || activeItem, activeItemChangeHandler }}>
-            <Styled.Aside ref={ref} position="left" id={id} isExpanded={isExpanded} isHovered={isHovered}>
-                <Styled.Nav isHovered={isHovered} isExpanded={isExpanded} onMouseEnter={openSidenav} onMouseLeave={closeSidenav}>
+            <Styled.Aside ref={ref} position="left" id={id} isExpanded={isExpanded} isHovered={isHovered} hideShadow={hideShadow}>
+                <Styled.Nav
+                    isHovered={isHovered}
+                    hideShadow={hideShadow}
+                    isExpanded={isExpanded}
+                    onMouseEnter={openSidenav}
+                    onMouseLeave={closeSidenav}
+                >
                     {children}
                     <ToggleSwitch id={`${id}-toggle`} isActive={isExpanded} onClick={isExpanded ? collapseSidenav : expandSidenav} />
                 </Styled.Nav>
@@ -50,3 +56,6 @@ SideNav.Group = NavGroup;
 SideNav.Context = SideNavContext;
 SideNav.displayName = 'SideNav';
 SideNav.Style = Styled.Aside;
+SideNav.defaultProps = {
+    hideShadow: false
+};