+++ packages/theme/src/core/singleSelect/types.ts
@@ -2,14 +2,55 @@ import { FontVariants } from '../font/types';
 
 export interface SingleSelectTheme {
     shadowColor: string;
+    variant: {
+        flat: {
+            height: string;
+            bgColor: {
+                default: string;
+                hovered: string;
+                pressed: string;
+                activated: string;
+                disabled: string;
+            };
+            labelColor: {
+                default: string;
+                hovered: string;
+                pressed: string;
+                activated: string;
+                disabled: string;
+                error: string;
+            };
+            valueColor: {
+                default: string;
+                hovered: string;
+                pressed: string;
+                activated: string;
+                disabled: string;
+                error: string;
+            };
+        };
+    };
+    options: {
+        borderRadius: {
+            flat: string;
+            filled: string;
+            outlined: string;
+        };
+    };
     option: {
         textVariant: {
             S: FontVariants;
             M: FontVariants;
         };
+        minHeight: {
+            flat: string;
+            filled: string;
+            outlined: string;
+        };
         bgColor: {
             default: string;
             hovered: string;
+            pressed: string;
             selected: string;
             error: string;
         };