+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.2](https://github.com/medly/medly-components/compare/@medly-components/utils@1.4.1...@medly-components/utils@1.4.2) (2020-09-10)
+
+**Note:** Version bump only for package @medly-components/utils
+
+
+
+
+
+## [1.4.1](https://github.com/medly/medly-components/compare/@medly-components/utils@1.4.0...@medly-components/utils@1.4.1) (2020-09-10)
+
+**Note:** Version bump only for package @medly-components/utils
+
+
+
+
+
 # [1.4.0](https://github.com/medly/medly-components/compare/@medly-components/utils@1.3.6...@medly-components/utils@1.4.0) (2020-09-08)